+++ resources/views/admin/order/view_orders.blade.php
@@ -70,7 +70,7 @@
                                             <td>{{ $order->customer->name }}</td>
                                             <td>{{ $order->created_at->toFormattedDateString() }}</td>
                                             <td>{{ $order->total_products }}</td>
-                                            <td>{{ $order->total }}</td>
+                                            <td><span>&#8358;</span>{{ $order->total }}</td>
                                             <td>{{ $order->payment_status }}</td>
                                             <td><span class="badge badge-warning">{{ $order->order_status }}</span></td>
 
@@ -82,14 +82,16 @@
 													btn-info">
                                                     <i class="fa fa-pencil-square-o" aria-hidden="true"></i>
                                                 </a>
-                                                <button class="btn btn-danger" type="button" onclick="deleteItem({{ $order->id }})">
-                                                    <i class="fa fa-trash" aria-hidden="true"></i>
-                                                </button>
-                                                <form id="delete-form-{{ $order->id }}" action="{{ route('admin.product.destroy', $order->id) }}" method="post"
-                                                      style="display:none;">
-                                                    @csrf
-                                                    @method('DELETE')
-                                                </form>
+                                                @if (Auth::user()->hasRole('admin'))
+                                                    <button class="btn btn-danger" type="button" onclick="deleteItem({{ $order->id }})">
+                                                        <i class="fa fa-trash" aria-hidden="true"></i>
+                                                    </button>
+                                                    <form id="delete-form-{{ $order->id }}" action="{{ route('admin.product.destroy', $order->id) }}" method="post"
+                                                        style="display:none;">
+                                                        @csrf
+                                                        @method('DELETE')
+                                                    </form>
+                                                @endif
                                             </td>
                                         </tr>
                                     @endforeach
